Complmento testes.

+++ src/main/java/br/com/paulo/store/services/ProductsService.java
@@ -15,25 +15,53 @@ public class ProductsService {
 	private ProductRepository prodRepo;
 	
 	public Product novoProduto(Product produto) {
+		if(produto == null) {
+			return produto;
+		}
+		
+		if(produto.getId() != null) {
+			Product produto_existente = prodRepo.getOne(produto.getId());
+			if(produto_existente != null) {
+				return null;
+			}
+		}
+		
 		Product produto_inserido = prodRepo.saveAndFlush(produto);
 		
 		return produto_inserido;
 	}
 	
 	public Product atualizarProduto(Product produto) {
+		if(produto == null) {
+			return null;
+		}
+		
+		if(produto.getId() == null) {
+			return null;
+		}
+		
+		Product produto_existente = prodRepo.findOne(produto.getId());
+		if(produto_existente == null) {
+			return null;
+		}
+		
 		Product produto_inserido = prodRepo.saveAndFlush(produto);
 		
 		return produto_inserido;
 	}
 	
 	public void deleteProduct(Product produto) {
-		prodRepo.delete(produto);
-		prodRepo.flush();
+		if(produto != null) {
+			prodRepo.delete(produto);
+			prodRepo.flush();
+		}
 	}
 	
 	public void deleteProduct(Long id) {
-		prodRepo.delete(id);
-		prodRepo.flush();
+		if(id != null) {
+			prodRepo.delete(id);
+			prodRepo.flush();
+		}
 	}
 	
 	public List<Product> listarProdutos() {
